==> PythonScript/1_DataCleaning.ipynb <==
## 1. Data Cleaning

### This script contains the following:
1. Importing libraries and data
2. Addressing missing values
3. Addressing duplicate values
4. Column dropping
5. Readdressing missing values
6. Exporting cleaned dataframes

### 1. Importing libraries and data

import pandas as pd
import numpy as np
import os

path = r'C:\Users\Neena Tilton\Dropbox\Projects\MinWage_Crime'

# Import Minimum wage by state dataframe.

df_mw = pd.read_csv(r"C:\Users\Neena Tilton\Dropbox\Projects\MinWage_Crime\02_Data\OriginalData\MinWage.csv", encoding='cp1252')

# Import crime by state dataframe. 

df_crime = pd.read_csv(os.path.join(path, '02_Data', 'OriginalData', 'crime.csv'))

df_mw.head()

df_crime.head()

df_mw.shape

df_crime.shape

### 2. Addressing Missing Values

# Check for null values in Min Wage df.

df_mw.isnull().sum()

Some missing value present, but only in columns that will be dropped. 

# Check for null values in crime df.

df_crime.isnull().sum()

Quite a bit of null values, but also mostly for columns that will be dropped. The only column that is of concern at present is the 'state_population' column. 

Address missing values in the 'state_population' column:

# Create subset where 'state_population' is null.

df_nan = df_crime[df_crime['state_population'].isnull() == True]
df_nan

The 16 of the 17 rows with NaN in the 'state_population' are of no issue since they are connected to FEDERAL prison count, not to a state. Only one row with the issue is row 746 for state of New York for 2015. We do need a value. The NaN value for this row will be addressed by replacing NaN with the average between 2014 and 2016. 

df_statepop = df_crime[df_crime['jurisdiction'] == 'NEW YORK']

df_statepop['state_population']

# Simple calculation of average between 2014 and 2016.

avg = (19746227+19836286)/2
avg

df_crime.at[746, 'state_population'] = avg

df_statepop_2 = df_crime[df_crime['jurisdiction'] == 'NEW YORK']
df_statepop_2[['jurisdiction', 'state_population']]

The rest of the NaN values will be ignored for now, will be dropped later. 

### 3. Addressing Duplicate Values

# Create subsets containing only rows that are duplicates of another row in both df.

df_mw_dups = df_mw[df_mw.duplicated()]
df_crime_dups = df_crime[df_crime.duplicated()]

df_mw_dups

df_crime_dups

Neither dups dataframe contained any rows, thus there are no duplicate rows in either of the dataframes.

### 4. Column Dropping

# Drop unneeded columns from Min Wage df.

df_mw.columns

df_mw = df_mw.drop(columns = [ 'CPI.Average','Department.Of.Labor.Uncleaned.Data',
       'Department.Of.Labor.Cleaned.Low.Value','Department.Of.Labor.Cleaned.Low.Value.2020.Dollars',
       'Department.Of.Labor.Cleaned.High.Value','Department.Of.Labor.Cleaned.High.Value.2020.Dollars', 
        'Footnote'], axis = 1)

df_mw.head()

df_mw.columns

# Drop unneeded columns from Crimes df.

df_crime.columns

df_crime = df_crime.drop(['includes_jails', 'crime_reporting_change', 
                          'crimes_estimated', 'rape_legacy','rape_revised', 
                           'agg_assault', 'property_crime_total','larceny', 
                          'vehicle_theft'], axis = 1)

df_crime.head()

df_crime.columns

### 5. Readdressing Missing Values

Ended up not dropping all the columns containing missing values. Need to re-address some of the null values in the crimes df. 

df_crime.isnull().sum()

df_nan = df_crime[df_crime['violent_crime_total'].isnull() == True]
df_nan

It is clear that crucial values were never reported for New York in 2015. Will treate each NaN value for this row the same way as 'state_population' earlier.

df_crime[df_crime['jurisdiction'] == 'NEW YORK']

# Impute values for 2015 by taking values from 2014 and 2016. 

violent_crime = (75398+74315)/2
violent_crime

murder = (617+629)/2
murder

robbery = (24045+22310)/2
robbery

burg = (50781+39850)/2
burg

# Replace NaN with imputed values.

df_crime.at[746, 'violent_crime_total'] = violent_crime

df_crime.at[746, 'murder_manslaughter'] = murder

df_crime.at[746, 'robbery'] = robbery

df_crime.at[746, 'burglary'] = burg

df_crime[df_crime['jurisdiction'] == 'NEW YORK']

df_crime.isnull().sum()

The remaining null values are linked to Federal prison information. Will leave as is. 

### 6. Exporting Cleaned df

df_mw.to_pickle(os.path.join(path, '02_Data', 'PreparedData', 'MinWage_clean.pkl'))

df_crime.to_pickle(os.path.join(path, '02_Data', 'PreparedData', 'crimes_clean.pkl'))
