==> PythonScript/2_DataWrangling.ipynb <==
## 2. Data Wrangling

### This script contains the following: 
1. Importing libraries and data
2. Renaming columns
3. Data consistency check
4. Data type check
5. Exporting data to pickle

### 1. Importing libraries and data

import pandas as pd
import numpy as np
import os

path = r'C:\Users\Neena Tilton\Dropbox\Projects\MinWage_Crime'

# Import dataframe of Minimum Wage

df_mw = pd.read_pickle(os.path.join(path, '02_Data', 'PreparedData', 'MinWage_clean.pkl'))

# Import dataframe of Crimes

df_crime = pd.read_pickle(os.path.join(path, '02_Data', 'PreparedData', 'crimes_clean.pkl'))

df_mw.head()

df_crime.head()

df_mw.columns

df_crime.columns

# Create a dictionary of old name to new name in mw_dict.

mw_dict = {'State.Minimum.Wage':'state_mw',
       'State.Minimum.Wage.2020.Dollars':'state_mw_2020',
        'Federal.Minimum.Wage':'fed_mw',
       'Federal.Minimum.Wage.2020.Dollars': 'fed_mw_2020',
        'Effective.Minimum.Wage': 'effective_mw',
       'Effective.Minimum.Wage.2020.Dollars':'effective_mw_2020'}

# Rename columns for Min Wage df using mw_dict.

df_mw.rename(columns = mw_dict, inplace = True)

df_mw.head()

# Create a dictionary of old name to new name in crime_dict.

crime_dict = {'jurisdiction':'state','violent_crime_total':'violen_crime', 
              'murder_manslaughter':'murder'}

df_crime.rename(columns = crime_dict, inplace = True)

df_crime.head()

### 3. Data consistency check

df_mw['State'].value_counts(dropna = False)

df_mw['Year'].value_counts(dropna = False)

df_crime['state'].value_counts(dropna = False)

df_crime['year'].value_counts(dropna = False)

Nothing out of ordinary as for data formatting for Year and State for both dataframes. 

### 4. Data type check

Here, any data type that take up too much memory will be updated to take up less space. 

df_mw.info()

df_mw.columns

# Change data type of Min Wage dataframe.

df_mw = df_mw.astype({'Year': 'int32', 
                'State': 'category',
                'state_mw': 'float32', 
                'state_mw_2020': 'float32', 
                'fed_mw': 'float32', 
                'fed_mw_2020': 'float32',
                'effective_mw': 'float32', 
                'effective_mw_2020': 'float32'})

df_mw.info()

Decreased memory usage from 179.0+ KB to 83.7 KB.

df_crime.info()

df_crime.columns

# Change data types of Crimes df.

df_crime = df_crime.astype({'state':'category',
                            'year':'int32', 
                            'prisoner_count':'float32', 
                            'state_population':'float32', 
                            'violen_crime':'float32',
                            'murder':'float32',  
                            'robbery':'float32', 
                            'burglary':'float32'})

df_crime.info()

Decreased memory usage from 51.1 KB to 25.7 KB.

Found typo in column name, change 'violen_crime' to 'violent_crime':

df_crime.rename(columns = {'violen_crime': 'violent_crime'}, inplace = True)

df_crime.columns

### 5. Export dataframes to pickle

df_mw.to_pickle(os.path.join(path, '02_Data', 'PreparedData', 'MinWage_wrangled.pkl'))

df_crime.to_pickle(os.path.join(path, '02_Data', 'PreparedData', 'crimes_wrangled.pkl'))
