==> PythonScript/3_Subsetting&Merging.ipynb <==
#### 3. Subsetting & Merging

### This script contains the following:
1. Importing libraries and data
2. Subsetting dataframes
3. Merging dataframes
4. Exporting merged dataframe

### 1. Importing libraries and data

import pandas as pd
import numpy as np
import os

path = r'C:\Users\Neena Tilton\Dropbox\Projects\MinWage_Crime'

# Import dataframe for Min Wage

df_mw = pd.read_pickle(os.path.join(path, '02_Data', 'PreparedData', 'MinWage_wrangled.pkl'))

# Import dataframe for Crime

df_crime = pd.read_pickle(os.path.join(path, '02_Data', 'PreparedData', 'crimes_wrangled.pkl'))

df_mw.head()

df_crime.head()

### 2. Subsetting Dataframe

After some consideration, decided that the rows for states named 'FEDERAL' is not needed of the Crimes df. Create a subset removing all 'FEDERAL' rows.

df_crime_2 = df_crime.loc[df_crime['state'] != 'FEDERAL']

df_crime_2['state'].value_counts(dropna = False)

df_crime_2.isnull().sum()

As for Min Wage dataframe, the 'Year' ranges between 1968 to 2020. The Crime data frame ranges between 2001 to 2016. Create a subset of Min Wage dataframe ranging only between 2001 to 2016. 

df_mw_2 = df_mw.loc[(df_mw['Year'] >= 2001) & (df_mw['Year'] <= 2016)]

df_mw_2['Year'].value_counts(dropna = False)

### 3. Merging Dataframes

The Crimes dataframe names for 'state' are in all caps. Update to be in proper title format to successfully merge the two dataframes. 

df_crime_2['state'] = df_crime_2['state'].str.title()

df_crime_2['state'].value_counts(dropna = False)

Now time to merge the Crimes df to Min Wage df.

# Before we do so, rename column names to match each other.

dict = {'state':'State', 'year':'Year'}

df_crime_2.rename(columns = dict, inplace = True)

df_crime_2.head()

df_merged = df_mw_2.merge(df_crime_2, on = ['State','Year'], how = 'inner', indicator = True)

df_merged.head()

df_merged['_merge'].value_counts()

df_merged.shape

# Now that we've confirmed that the merged df is a full merge, remove the '_merge' flag column.

df_merged = df_merged.drop(columns = ['_merge'])

df_merged.shape

df_merged.columns

### 4. Export merged dataframe

df_merged.to_pickle(os.path.join(path, '02_Data','PreparedData', 'df_merged.pkl'))
