==> PythonScript/5_DescriptiveStatistics.ipynb <==
## 5. Descriptive Statistics

### This script contains the following:
1. Importing libraries and data
2. Running descriptive statistics analysis
3. Exporting tables to csv files

### 1. Importing libraries and data

import pandas as pd
import numpy as np
import os

path = r'C:\Users\Neena Tilton\Dropbox\Projects\MinWage_Crime'

df = pd.read_pickle(os.path.join(path, '02_Data', 'PreparedData', 'df_newvar.pkl'))

df.head()

### 2. Running descriptive statistics analysis

df_stats = df.describe()
df_stats

This is a great opportunity to once again all columns have complete values and within reasonable range of expectations or find anomalies. Looks like all columns contain 800 enteries, indicating no empty rows present. For 'Year' the min value is 2001 and max value is 2016, no anomolies here. All minimum wage data (marked with 'mw' in the column name) are within reasonable ranges for minimum wages. Within the columns with the types of crimes, the only suspect value is for the 'murder' column, where the minimum is 5. This is wholly possible, thus will not treat as outlier, however, am curious to see which state in which year had such low rate of such type of crime. 

### 3. Exporting stats table to csv file

# Save as csv file for possible use later.

df_stats.to_csv(os.path.join(path,'02_Data', 'PreparedData', 'df_descptive.csv'))
